+++ circuits/cpp/barretenberg/src/aztec/stdlib/primitives/bigfield/bigfield_impl.hpp
@@ -86,7 +86,7 @@ bigfield<C, T>::bigfield(const field_t<C>& low_bits, const field_t<C>& high_bits
         std::vector<uint32_t> high_accumulator;
         if constexpr (C::type == waffle::PLOOKUP) {
             ASSERT(NUM_LIMB_BITS % 2 == 0); // required for one of the intermediate sums giving the third limb
-            high_accumulator = context->decompose_into_default_range(high_bits.witness_index, num_high_limb_bits);
+            high_accumulator = context->decompose_into_default_range(high_bits.witness_index, static_cast<uint32_t>(num_high_limb_bits));
             size_t mid_index = (NUM_LIMB_BITS / waffle::PlookupComposer::DEFAULT_PLOOKUP_RANGE_BITNUM) / 2 - 1;
             limb_2.witness_index = high_accumulator[mid_index];
             limb_3 = (high_bits - limb_2) * shift_right_1;
